+++ x/feemarket/keeper/eip1559.go
@@ -1,6 +1,7 @@
 package keeper
 
 import (
+	ethparams "github.com/ethereum/go-ethereum/params"
 	"math/big"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -12,32 +13,19 @@ import (
 // CalculateBaseFee calculates the base fee for the current block. This is only calculated once per
 // block during BeginBlock. If the NoBaseFee parameter is enabled or below activation height, this function returns nil.
 // NOTE: This code is inspired from the go-ethereum EIP1559 implementation and adapted to Cosmos SDK-based
-// chains. For the canonical code refer to: https://github.com/ethereum/go-ethereum/blob/master/consensus/misc/eip1559.go
+// chains. For the canonical code refer to: https://github.com/ethereum/go-ethereum/blob/v1.10.26/consensus/misc/eip1559.go
 func (k Keeper) CalculateBaseFee(ctx sdk.Context) *big.Int {
 	params := k.GetParams(ctx)
 
 	// Ignore the calculation if not enabled
-	if !params.IsBaseFeeEnabled(ctx.BlockHeight()) {
+	if params.NoBaseFee {
 		return nil
 	}
 
 	consParams := ctx.ConsensusParams()
 
-	// If the current block is the first EIP-1559 block, return the base fee
-	// defined in the parameters (DefaultBaseFee if it hasn't been changed by
-	// governance).
-	if ctx.BlockHeight() == params.EnableHeight {
-		return params.BaseFee.BigInt()
-	}
-
 	// get the block gas used and the base fee values for the parent block.
-	// NOTE: this is not the parent's base fee but the current block's base fee,
-	// as it is retrieved from the transient store, which is committed to the
-	// persistent KVStore after EndBlock (ABCI Commit).
 	parentBaseFee := params.BaseFee.BigInt()
-	if parentBaseFee == nil {
-		return nil
-	}
 
 	parentGasUsed := k.GetBlockGasWanted(ctx)
 
@@ -50,13 +38,13 @@ func (k Keeper) CalculateBaseFee(ctx sdk.Context) *big.Int {
 
 	// CONTRACT: ElasticityMultiplier cannot be 0 as it's checked in the params
 	// validation
-	parentGasTargetBig := new(big.Int).Div(gasLimit, new(big.Int).SetUint64(uint64(params.ElasticityMultiplier)))
+	parentGasTargetBig := new(big.Int).Div(gasLimit, new(big.Int).SetUint64(uint64(ethparams.ElasticityMultiplier)))
 	if !parentGasTargetBig.IsUint64() {
 		return nil
 	}
 
 	parentGasTarget := parentGasTargetBig.Uint64()
-	baseFeeChangeDenominator := new(big.Int).SetUint64(uint64(params.BaseFeeChangeDenominator))
+	baseFeeChangeDenominator := new(big.Int).SetUint64(uint64(ethparams.BaseFeeChangeDenominator))
 
 	// If the parent gasUsed is the same as the target, the baseFee remains
 	// unchanged.