+++ x/feemarket/keeper/params_test.go
@@ -34,22 +34,10 @@ func (suite *KeeperTestSuite) TestSetGetParams() {
 			true,
 		},
 		{
-			"success - Check ElasticityMultiplier is set to 3 and can be retrieved correctly",
-			func() interface{} {
-				params.ElasticityMultiplier = 3
-				err := suite.app.FeeMarketKeeper.SetParams(suite.ctx, params)
-				suite.Require().NoError(err)
-				return params.ElasticityMultiplier
-			},
-			func() interface{} {
-				return suite.app.FeeMarketKeeper.GetParams(suite.ctx).ElasticityMultiplier
-			},
-			true,
-		},
-		{
 			"success - Check BaseFeeEnabled is computed with its default params and can be retrieved correctly",
 			func() interface{} {
-				err := suite.app.FeeMarketKeeper.SetParams(suite.ctx, types.DefaultParams())
+				params.NoBaseFee = false
+				err := suite.app.FeeMarketKeeper.SetParams(suite.ctx, params)
 				suite.Require().NoError(err)
 				return true
 			},
@@ -62,7 +50,6 @@ func (suite *KeeperTestSuite) TestSetGetParams() {
 			"success - Check BaseFeeEnabled is computed with alternate params and can be retrieved correctly",
 			func() interface{} {
 				params.NoBaseFee = true
-				params.EnableHeight = 5
 				err := suite.app.FeeMarketKeeper.SetParams(suite.ctx, params)
 				suite.Require().NoError(err)
 				return true