+++ x/feemarket/types/params.go
@@ -6,7 +6,7 @@ import (
 	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
-	"github.com/ethereum/go-ethereum/params"
+	ethparams "github.com/ethereum/go-ethereum/params"
 )
 
 var (
@@ -14,22 +14,17 @@ var (
 	DefaultMinGasMultiplier = sdk.NewDecWithPrec(50, 2)
 	// DefaultMinGasPrice is 0 (i.e disabled)
 	DefaultMinGasPrice = sdk.ZeroDec()
-	// DefaultEnableHeight is 0 (i.e disabled)
-	DefaultEnableHeight = int64(0)
 	// DefaultNoBaseFee is false
 	DefaultNoBaseFee = false
 )
 
 // Parameter keys
 var (
-	ParamsKey                             = []byte("Params")
-	ParamStoreKeyNoBaseFee                = []byte("NoBaseFee")
-	ParamStoreKeyBaseFeeChangeDenominator = []byte("BaseFeeChangeDenominator")
-	ParamStoreKeyElasticityMultiplier     = []byte("ElasticityMultiplier")
-	ParamStoreKeyBaseFee                  = []byte("BaseFee")
-	ParamStoreKeyEnableHeight             = []byte("EnableHeight")
-	ParamStoreKeyMinGasPrice              = []byte("MinGasPrice")
-	ParamStoreKeyMinGasMultiplier         = []byte("MinGasMultiplier")
+	ParamsKey                     = []byte("Params")
+	ParamStoreKeyNoBaseFee        = []byte("NoBaseFee")
+	ParamStoreKeyBaseFee          = []byte("BaseFee")
+	ParamStoreKeyMinGasPrice      = []byte("MinGasPrice")
+	ParamStoreKeyMinGasMultiplier = []byte("MinGasMultiplier")
 )
 
 // ParamKeyTable returns the parameter key table.
@@ -41,10 +36,7 @@ func ParamKeyTable() paramtypes.KeyTable {
 func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
 	return paramtypes.ParamSetPairs{
 		paramtypes.NewParamSetPair(ParamStoreKeyNoBaseFee, &p.NoBaseFee, validateBool),
-		paramtypes.NewParamSetPair(ParamStoreKeyBaseFeeChangeDenominator, &p.BaseFeeChangeDenominator, validateBaseFeeChangeDenominator),
-		paramtypes.NewParamSetPair(ParamStoreKeyElasticityMultiplier, &p.ElasticityMultiplier, validateElasticityMultiplier),
 		paramtypes.NewParamSetPair(ParamStoreKeyBaseFee, &p.BaseFee, validateBaseFee),
-		paramtypes.NewParamSetPair(ParamStoreKeyEnableHeight, &p.EnableHeight, validateEnableHeight),
 		paramtypes.NewParamSetPair(ParamStoreKeyMinGasPrice, &p.MinGasPrice, validateMinGasPrice),
 		paramtypes.NewParamSetPair(ParamStoreKeyMinGasMultiplier, &p.MinGasMultiplier, validateMinGasPrice),
 	}
@@ -53,49 +45,34 @@ func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
 // NewParams creates a new Params instance
 func NewParams(
 	noBaseFee bool,
-	baseFeeChangeDenom,
-	elasticityMultiplier uint32,
 	baseFee uint64,
-	enableHeight int64,
 	minGasPrice sdk.Dec,
 	minGasPriceMultiplier sdk.Dec,
 ) Params {
 	return Params{
-		NoBaseFee:                noBaseFee,
-		BaseFeeChangeDenominator: baseFeeChangeDenom,
-		ElasticityMultiplier:     elasticityMultiplier,
-		BaseFee:                  sdkmath.NewIntFromUint64(baseFee),
-		EnableHeight:             enableHeight,
-		MinGasPrice:              minGasPrice,
-		MinGasMultiplier:         minGasPriceMultiplier,
+		NoBaseFee:        noBaseFee,
+		BaseFee:          sdkmath.NewIntFromUint64(baseFee),
+		MinGasPrice:      minGasPrice,
+		MinGasMultiplier: minGasPriceMultiplier,
 	}
 }
 
 // DefaultParams returns default evm parameters
 func DefaultParams() Params {
 	return Params{
-		NoBaseFee:                DefaultNoBaseFee,
-		BaseFeeChangeDenominator: params.BaseFeeChangeDenominator,
-		ElasticityMultiplier:     params.ElasticityMultiplier,
-		BaseFee:                  sdkmath.NewIntFromUint64(params.InitialBaseFee),
-		EnableHeight:             DefaultEnableHeight,
-		MinGasPrice:              DefaultMinGasPrice,
-		MinGasMultiplier:         DefaultMinGasMultiplier,
+		NoBaseFee:        DefaultNoBaseFee,
+		BaseFee:          sdkmath.NewIntFromUint64(ethparams.InitialBaseFee),
+		MinGasPrice:      DefaultMinGasPrice,
+		MinGasMultiplier: DefaultMinGasMultiplier,
 	}
 }
 
 // Validate performs basic validation on fee market parameters.
 func (p Params) Validate() error {
-	if p.BaseFeeChangeDenominator == 0 {
-		return fmt.Errorf("base fee change denominator cannot be 0")
-	}
-
-	if p.BaseFee.IsNegative() {
-		return fmt.Errorf("initial base fee cannot be negative: %s", p.BaseFee)
-	}
-
-	if p.EnableHeight < 0 {
-		return fmt.Errorf("enable height cannot be negative: %d", p.EnableHeight)
+	if p.BaseFee.IsNil() {
+		return fmt.Errorf("base fee cannot be nil")
+	} else if p.BaseFee.IsNegative() {
+		return fmt.Errorf("base fee cannot be negative: %s", p.BaseFee)
 	}
 
 	if err := validateMinGasMultiplier(p.MinGasMultiplier); err != nil {
@@ -111,10 +88,6 @@ func validateBool(i interface{}) error {
 		return fmt.Errorf("invalid parameter type: %T", i)
 	}
 	return nil
-}
-
-func (p *Params) IsBaseFeeEnabled(height int64) bool {
-	return !p.NoBaseFee && height >= p.EnableHeight
 }
 
 func validateMinGasPrice(i interface{}) error {
@@ -135,27 +108,6 @@ func validateMinGasPrice(i interface{}) error {
 	return nil
 }
 
-func validateBaseFeeChangeDenominator(i interface{}) error {
-	value, ok := i.(uint32)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-
-	if value == 0 {
-		return fmt.Errorf("base fee change denominator cannot be 0")
-	}
-
-	return nil
-}
-
-func validateElasticityMultiplier(i interface{}) error {
-	_, ok := i.(uint32)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-	return nil
-}
-
 func validateBaseFee(i interface{}) error {
 	value, ok := i.(sdkmath.Int)
 	if !ok {
@@ -164,19 +116,6 @@ func validateBaseFee(i interface{}) error {
 
 	if value.IsNegative() {
 		return fmt.Errorf("base fee cannot be negative")
-	}
-
-	return nil
-}
-
-func validateEnableHeight(i interface{}) error {
-	value, ok := i.(int64)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-
-	if value < 0 {
-		return fmt.Errorf("enable height cannot be negative: %d", value)
 	}
 
 	return nil